+++ app/javascript/packs/application.js
@@ -27,10 +27,12 @@ import "bootstrap";
 
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
-import {initMapBox} from '../components/initMapBox';
+import { initMapbox } from '../plugins/init_mapbox';
+
 document.addEventListener('turbolinks:load', () => {
+
   // Call your functions here, e.g:
   // initSelect2();
-  initMapBox();
+  initMapbox();
 });
