+++ app/views/stalls/index.html.erb
@@ -31,9 +31,7 @@
     <div id="map" 
     style="width: 100%; height: 600px;"
      data-markers="<%= @markers.to_json %>"
-     data-mapbox-api-key="<%= ENV[''] %>">
+     data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>">
      </div> 
 </div>
-<%#pk.eyJ1IjoiY2FtZWxpYTY3OCIsImEiOiJja3ZtajJ4eXMwNm1sMnBxaHl6N3phNmJlIn0.2t7uQsHWjv1M2Kq1NPnAJQ%>
-   <%# <img id="map" src='https://api.mapbox.com/styles/v1/mapbox/streets-v11/static/2.3381,48.8693,10.45,0,0/600x1000?access_token=pk.eyJ1IjoiY2FtZWxpYTY3OCIsImEiOiJja3ZtajJ4eXMwNm1sMnBxaHl6N3phNmJlIn0.2t7uQsHWjv1M2Kq1NPnAJQ'>
-  </div> %>
+
