+++ package.json
@@ -10,10 +10,11 @@
     "jquery": "^3.6.0",
     "mapbox-gl": "^2.6.1",
     "popper.js": "^1.16.1",
-    "turbolinks": "^5.2.0"
+    "turbolinks": "^5.2.0",
+    "webpack-dev-server": "3.11.2"
   },
   "version": "0.1.0",
   "devDependencies": {
-    "webpack-dev-server": "^4.5.0"
+    "@webpack-cli/serve": "^1.6.0"
   }
 }
